feat: move theme state to app component
